refactor: Init new form response lib
Starting a new lib for form responses. Currently only moved the code from formbuilder-utils.
Will now start re-writing it to be a lot cleaner and easy to use.
